requirements.txt
amqp==5.3.1
asgiref==3.11.1
attrs==25.4.0
billiard==4.2.4
brotli==1.2.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cssselect2==0.9.0
Django==6.0.3
django-cors-headers==4.9.0
django-filter==25.2
django-ratelimit==4.1.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
drf-spectacular==0.29.0
fonttools==4.61.1
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kombu==5.6.2
librt==0.8.1
mypy==1.19.1
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
pillow==12.1.1
pluggy==1.6.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==3.0
pydyf==0.12.1
Pygments==2.19.2
pyphen==0.17.2
pytest==9.0.2
pytest-django==4.12.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
redis==7.2.1
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
ruff==0.15.5
six==1.17.0
sqlparse==0.5.5
tinycss2==1.5.1
tinyhtml5==2.1.0
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
weasyprint==68.1
httpx==0.28.1
PyPDF2==3.0.1
python-docx==1.1.2
webencodings==0.5.1
zopfli==0.4.1
